Rename main to main1
The first code. I found a basic code in internet and I adapted for this proyect. In this moment I only have spare parts of my printer 3d so the code is for arduino mega ,ramps 1.4 and lcd reprap 2004. Also is necesary use 2 basic buttons with 2 resistors for use the menu.